requirements.txt
fastapi==0.111.0
uvicorn==0.22.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.10
openai>=1.0.0
faiss-cpu==1.12.0
pydantic>=2.7.4
python-dotenv==1.0.0
langsmith>=0.1.99
pytest==7.4.0
pytest-asyncio==0.22.0
networkx>=3.0
langchain-google-genai>=0.0.1
